Move fizz specific part of the client in quic/fizz/client
