+++ quic/client/QuicClientTransport.h
@@ -14,9 +14,9 @@
 #include <folly/io/async/AsyncUDPSocket.h>
 #include <folly/net/NetOps.h>
 #include <quic/api/QuicTransportBase.h>
-#include <quic/client/handshake/QuicPskCache.h>
 #include <quic/client/state/ClientStateMachine.h>
 #include <quic/common/BufUtil.h>
+#include <quic/fizz/client/handshake/QuicPskCache.h>
 
 namespace quic {
