+++ quic/client/handshake/ClientHandshake.cpp
@@ -9,8 +9,8 @@
 #include <quic/client/handshake/ClientHandshake.h>
 
 #include <quic/client/handshake/ClientTransportParametersExtension.h>
-#include <quic/client/handshake/QuicPskCache.h>
 #include <quic/client/state/ClientStateMachine.h>
+#include <quic/fizz/client/handshake/QuicPskCache.h>
 #include <quic/state/QuicStreamFunctions.h>
 
 namespace quic {
