+++ quic/fizz/CMakeLists.txt
@@ -3,4 +3,5 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+add_subdirectory(client)
 add_subdirectory(handshake)
