+++ quic/samples/echo/EchoClient.h
@@ -18,9 +18,9 @@
 
 #include <quic/api/QuicSocket.h>
 #include <quic/client/QuicClientTransport.h>
-#include <quic/client/handshake/FizzClientQuicHandshakeContext.h>
 #include <quic/common/BufUtil.h>
 #include <quic/common/test/TestUtils.h>
+#include <quic/fizz/client/handshake/FizzClientQuicHandshakeContext.h>
 
 namespace quic {
 namespace samples {
